+++ Tests/BaseTestForNUnit.cs
@@ -15,8 +15,7 @@ namespace Tests
         {
             Flow.Setup(
                 GetType(),
-                TestContext.CurrentContext.Test.FullName,
-                () => TestContext.CurrentContext.Test.FullName);
+                TestContext.CurrentContext.Test.FullName);
         }
 
         [TearDown]